requirements.txt
streamlit
torch
torchvision
pillow
google-genai
